manually create redirects for all xml files
This is a workaround for integration with the discourse forum.

Discourse normalizes links to be lower case for some reason, so this adds redirects for the lowercased
version of every xml file to the upper case version.

See this thread on discourse Meta: https://meta.discourse.org/t/discourse-embed-normalizes-urls-to-be-lowercase/295747/2.

The better way to do this would be via a regex redirect, however Cloudflare pages does not support this.